Remove one of r12797 casts as the issue will be fixed upon next trunk merge

+++ branches/5.2.0/src/psimage.c
@@ -419,7 +419,7 @@ int GMT_psimage (void *V_API, int mode, void *args)
 		}
 		else if (I->header->n_bands == 4) { /* RGBA image, with a color map */
 			if (!Ctrl->G.active) has_trans = find_unique_color (GMT, I->data, (int)I->header->nm, &r, &g, &b);
-			for (j = n = 0; j < 4 * (int)I->header->nm; j++) { /* Reduce image from 32- to 24-bit */
+			for (j = n = 0; j < 4 * I->header->nm; j++) { /* Reduce image from 32- to 24-bit */
 				if (has_trans && I->data[j+3] == 0)
 					I->data[n++] = r, I->data[n++] = g, I->data[n++] = b, j+=3;
 				else
